+++ src/Rotatable.java
@@ -2,6 +2,6 @@
  * Interface for an Object which is both coordinate and direction dependent.
  */
 public interface Rotatable extends Mappable {
-    void setDirection(double direction);
-    double getDirection();
+    void setDirection(int direction);
+    int getDirection();
 }
